% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvUploadModule.R
\name{csvUploadServer}
\alias{csvUploadServer}
\title{Server Function for CSV Upload Module}
\usage{
csvUploadServer(id, vars_title = "Available Variables")
}
\arguments{
\item{id}{A character string that matches the ID used in \code{csvUploadUI}.}

\item{vars_title}{A character string for the title of the variables table.}
}
\value{
A reactive expression containing the uploaded data.
}
\description{
This function defines the server logic for the CSV upload module.
}
\examples{
server <- function(input, output, session) {
  csvUploadServer("my_data", "My Variables")
}

}
