% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvUploadModule.R
\name{csvUploadUI}
\alias{csvUploadUI}
\title{UI Function for CSV Upload Module}
\usage{
csvUploadUI(id)
}
\arguments{
\item{id}{A character string that defines the namespace for the module.}
}
\value{
A list containing two elements:
\itemize{
\item \code{input}: The file input UI for uploading a CSV file.
\item \code{output}: The UI for displaying the variables table.
}
}
\description{
This function creates the UI elements for the CSV upload module.
}
\examples{
csvUploadUI("my_data")

}
