% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{format_citation}
\alias{format_citation}
\title{Format Citation}
\usage{
format_citation(cit)
}
\arguments{
\item{cit}{A citation object obtained from \code{utils::citation()}.}
}
\value{
A character string containing the formatted citation.
}
\description{
This function formats a citation object into a format ready for use in 'shiny'
applications
}
\examples{
format_citation(utils::citation("base"))

}
