% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresModule.R
\name{postgresUI}
\alias{postgresUI}
\title{Create UI elements for Postgres Shiny Module}
\usage{
postgresUI(id)
}
\arguments{
\item{id}{A character string that uniquely identifies this module instance}
}
\value{
A list containing three UI elements:
\item{submit}{An action button for submitting data to database}
\item{table}{A DT output for displaying the database data}
\item{download}{A download button for exporting database data to csv}
}
\description{
This function generates the UI components for the Postgres Shiny module,
including a submit button, a data table, and a download button.
}
\examples{
shiny::fluidPage(
  postgresUI("postgres_module")$submit,
  postgresUI("postgres_module")$table,
  postgresUI("postgres_module")$download
  )
}
