% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_to_server.R
\name{text_to_list}
\alias{text_to_list}
\title{Convert Text Input to List}
\usage{
text_to_list(text_input)
}
\arguments{
\item{text_input}{A text representation of a list.}
}
\value{
A list parsed from the input string.
}
\description{
Convert Text Input to List
}
\examples{
# Create a named list
text_to_list("'one' = 1, 'two' = 2, 'three' = 3")

# Create a list of vectors
text_to_list("c('x1', 'x2'), c('x3', 'x4')")

}
