% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_to_server.R
\name{text_to_vector}
\alias{text_to_vector}
\title{Convert Text Input to Vector}
\usage{
text_to_vector(text_input)
}
\arguments{
\item{text_input}{A string input to be converted to a vector.}
}
\value{
A vector parsed from the input string.
}
\description{
The goal of this function is to take text input and convert it
to an R vector.
}
\examples{

text_to_vector("1,2,3,4,5")

text_to_vector("1:5")

text_to_vector("rep(1:5, times = 2)")

text_to_vector("seq(1,10,2)")

}
